+++ console/Packages/com.cronyx.console/CHANGELOG.md
@@ -7,9 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added a public overload for `DeveloperConsole.RegisterCommand` that takes an `IConsoleCommand` instance.
+- Added `VerbCommand`, a new `IConsoleCommand` implementation, with support for verb commands, as in the case of `git push` or `git pull`
+    - Allows for multiple subcommands to be called from one single root command by specifying *verbs*
+    - Use the `RegisterVerb` and its overloads to register verb implementations
 
 ### Changed
 - Made the formatting of the `help` command more rigorous, supporting multi-line command descriptions and coloring built-in commands blue.
+- Renamed `DeveloperConsole.RegisterCommand(string, Action<string>, string, string)` overload to `DeveloperConsole.RegisterCommandManual` to avoid ambiguity with `DeveloperConsole.RegisterCommand<T0>`.
+
+### Fixed
+- Fixed target exception when calling automatically parsed commands when using non-static methods.
 
 ## [1.1.1] - 2021-03-23
 ### Changed